feat(tests): migrate test runner from tsx scripts to Vitest
Closes #124.

Replaces ad-hoc top-level tsx scripts with proper Vitest describe/it/expect
suites. All 15 assertions preserved; coverage reporting enabled (no threshold
yet — baseline is lines: 74.3%, measured after first run per issue spec).

Changes:
- Add vitest + @vitest/coverage-v8 to devDependencies; remove tsx
- Add vitest.config.ts with v8 coverage provider (no threshold)
- Rename tests/**/*.ts → *.test.ts; convert imperative code to describe/it
- Consolidate 4 CI test jobs (test-contracts/goldens/e2e/learnings) into one
  "Tests (Vitest)" job running `pnpm test`
- Update quality-gate aggregator to reference single test job

Coverage baseline (lines: 74.3%) documented; threshold to be added in follow-up
once team has agreed on target per-module.
