+++ package.json
@@ -20,17 +20,16 @@
   "packageManager": "pnpm@9.0.0",
   "scripts": {
     "check:governance": "bash scripts/check-governance.sh",
-    "test:contracts": "tsx tests/contracts/test-contracts.ts",
-    "test:goldens": "tsx tests/goldens/test-goldens.ts",
-    "test:e2e": "tsx tests/e2e/test-e2e.ts",
-    "test:learnings": "tsx tests/learnings/test-learnings.ts",
-    "test": "pnpm run test:contracts && pnpm run test:goldens && pnpm run test:e2e && pnpm run test:learnings",
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage",
+    "test:watch": "vitest",
     "test:all": "pnpm run check:governance && pnpm run test"
   },
   "devDependencies": {
     "@types/node": "^20.14.0",
-    "tsx": "^4.19.0",
-    "typescript": "^5.5.0"
+    "@vitest/coverage-v8": "^4.1.7",
+    "typescript": "^5.5.0",
+    "vitest": "^4.1.7"
   },
   "dependencies": {
     "ajv": "^8.20.0",
